Add softDeleted and hardDeleted array fields to Appointment mongoose model
the softDeleted array field holds the array of sender and/or receiver that has deleted the message. The hardDeleted holds the array of sender or receiver that first deleted the message from his/her trash. During hard delete, the array is checked, if not empty the message is hard deleted otherwise the user ObjectId is kept in the array