[Bug Fix]Patient list crash with error accessing undefined 'about' field
Crash if the Doctor tries to access the patients list and some have not entered their 'about me' this throws the error because the 'about'
field is an entry on the 'BioData' field object. Also the user full name is returning undefined because it is usually accessed by the virtual field on name
i.e. name.full. This virtual field is inactive as the info is passed in through a session"

The whole story up is wrong. It crashes because it tries to access an undefined field. The data was passed in as a local initially with every field present but as the data is now passed in from the session, only few fields are defined in the patients Objects array in the doctors profile and do not comform to the initial field names.